Kill daml-foundations directory (#2219)
The only stuff left in there are design documents for contract keys and
the time model. The former is being superceded by making contract keys part
of the ledger model. The latter is out of date and will be replaced by
adding a time model to the ledger model as well.