Compute last seen location in speedy from stack trace
Currently, we use as the last seen location literally the last location
annotation the interpreter has seen. This is inaccurate since it does not take
into account that we could already have left the scope of this annotation.

After this PR we use the top of the stack trace we maintain internally as the
last location. This is more accurate since it takes leaving scopes into
account.