Fix async exception handling in grpc-haskell
This replaces the call to mask_ that I added in clientRequest as a
rudamentary fix by a proper fix that fixes the way async exceptions
are handled and should hopefully be more reasonable to upstream:

gRPC-haskell had a lot of cases where they did something like

```
do x <- allocateResource
   f x `finally` cleanup x
```

That breaks if you get an exception after allocating the resource but
before the exception handler is installed by finally. I replaced all
of those instances by `bracket` where possible and one explicit call
to `mask` where it is not possible.

There is a bit of boilerplate here since the resource allocation
returns an `Either` but given that there doesn’t seem to be an
existing utils module where I could put this and I don’t want to
depend on `extra` for things like `whenRight` (since I want to
upstream this), I’ve just kept the boilerplate for now.