Instruct GHC to use unicode output while capturing warning
The reason for this change is that code actions are generated by parsing
the output messages by GHC. By setting the useUnicode setting to always
be true, we can simplify parsing of these messages by not having to deal
with both unicode and non-unicode error messages and warnings.

One code action that relies on the error messages using unicode is the
'suggest to rename' code action.