ledger-api-test-tool: Sort the tests afterwards for better concurrency.
Turns out that Akka Streams tries to maintain order at the expense of
running more things, which is not ideal. This is a hack to let us use
`.mapAsyncUnordered` and then re-sort into the original order at the
end.

Ideally `.mapAsync` would detect when subsequent futures have finished
and start running more, even if the first hasn't finished yet.