ledger-api-test-tool: Allocate participants and parties declaratively.
Part of a solution to #3380.

Rather than allocating participants and parties line by line, instruct
the test to construct a number up-front.

This will make it easier to serve multi-participant ledgers, as we can
make sure we wait for all parties to be available on all participants
before starting the test. Once we can do that, tests will hopefully be
much less flaky on ledgers other than the Sandbox.