Fix potentical race in Dispatcher
Instead of using the value of the state at the time of subscription,
we need to call getHead explicitly.
Otherwise it could happen that the state gets updated between reading it
and completing the subscription, which means the stream would miss an update
of head.

This also removes the type parameter of SignalDispatcher again, as it is best
that this class doesn't even know the data type involved.