Relax restrictions on contract ids
Consider the following DAML code

```
template FooG a
  with
    p : Party
    a : a
  where
    signatory p

template instance Foo = FooG Decimal

template Boo
  with
    p : Party
  where
    signatory p

    controller p can
      nonconsuming ArchiveFoo: ()
        with
          fooCid : ContractId Foo
        do
          archive fooCid
```

This will generate an ArchiveFoo record with a field of type
`ContractId (FooG Decimal)` so no monomorphization takes place. This
is valid and is not rejected by the LF typechecker. However, currently
the validation logic here will reject this making it impossible to
call this choice.

This PR relaxes this restriction so that it is possible to call this choice.