Fix unreleased.sh
The usage of `git log` to avoid printing the diff of each commit turned
out to cause the commit to be interpreted as a revision range reachable
from said commit, which breaks unreleased.sh. This commit reverts to
using `git show` but with `--quiet` to avoid printing the diff, as
originally planned.